requirements.txt
tensorflow-cpu>=2.16.0,<2.22
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
tqdm>=4.65.0
streamlit>=1.30.0
huggingface_hub>=0.20.0
